requirements.txt
replicate>=0.25.0
httpx>=0.27.0
python-dotenv>=1.0.0
